Major programming update
Added shootable target for 2nd room. Player dies when they fall on spikes and will be promted to restart or quit. Player can now win the game. Added three new widgets: Timer, Win, Loss. Timer shows the player the countdown they have in order to complete the level. If they do not complete it in the alotted time they will fail. Win acts as a win screen once the player reaches the top. Loss is a screen that shows when the player touches the spikes or runs out of time. Added the levels GameFail and GameWin to show UI Widgets.